troca de chave para login + melhorias no cadastro de atividades curso de extensao

+++ src/main/java/com/ftunicamp/tcc/controllers/response/AtividadeResponse.java
@@ -12,7 +12,7 @@ public class AtividadeResponse implements Serializable {
 
     private Long id;
     private LocalDate dataCriacao;
-    private int prazo;
+    private long prazo;
     private String projeto;
 
 }
